functions/requirements.txt
azure-functions>=1.19.0
azure-identity>=1.17.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=32.0.0
azure-monitor-query>=1.3.0
azure-keyvault-secrets>=4.8.0
httpx>=0.27.0
pydantic>=2.7.0
python-dotenv>=1.0.0
structlog>=24.0.0
tenacity>=8.3.0
azure-cosmos>=4.7.0
